requirements.txt
# TAIP-ERT-Pipeline 相依套件
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.5.0
pygimli>=1.2.0
pyyaml>=6.0
# ftplib 和 pathlib 是 Python 標準庫，不需要額外安裝
plotly>=5.10.0
scikit-learn>=1.0.0
pandas>=1.3.0 